+++ src/components/package/edit-custom/custom-package-edit.js
@@ -5,9 +5,7 @@ import {
   useCallback,
 } from 'react';
 import PropTypes from 'prop-types';
-import {
-  Form
-} from 'react-final-form';
+import { Form } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
 import createFocusDecorator from 'final-form-focus';
 import {
@@ -19,7 +17,6 @@ import { useStripes } from '@folio/stripes/core';
 import {
   Button,
   Headline,
-  PaneFooter,
 } from '@folio/stripes/components';
 
 import {
@@ -38,6 +35,7 @@ import HoldingStatus from '../show/components/holding-status';
 import EditPackageInformation from '../edit/components/edit-package-information';
 import EditPackageSettings from '../edit/components/edit-package-settings';
 import EditCoverageSettings from '../edit/components/edit-coverage-settings';
+import EditPaneFooter from '../edit/components/edit-pane-footer';
 import {
   useStateCallback,
   useSectionToggle,
@@ -190,42 +188,18 @@ const CustomPackageEdit = ({
   };
 
   const getFooter = (pristine, reset) => {
-    const cancelButton = (
-      <Button
-        data-test-eholdings-package-edit-cancel-button
-        buttonStyle="default mega"
-        disabled={model.update.isPending || pristine}
-        onClick={reset}
-        marginBottom0
-      >
-        <FormattedMessage id="stripes-components.cancel" />
-      </Button>
-    );
-
-    const saveButton = (
-      <Button
-        buttonStyle="primary mega"
-        data-test-eholdings-package-save-button
-        disabled={model.update.isPending || pristine}
-        marginBottom0
-        type="submit"
-      >
-        <FormattedMessage id="stripes-components.saveAndClose" />
-      </Button>
-    );
-
     return (
-      <PaneFooter
-        renderStart={cancelButton}
-        renderEnd={saveButton}
+      <EditPaneFooter
+        disabled={model.update.isPending || pristine}
+        reset={reset}
       />
     );
   };
 
   return (
     <KeyShortcutsWrapper
-      formRef={editFormRef.current}
       toggleAllSections={toggleAllSections}
+      formRef={editFormRef.current}
     >
       <Form
         onSubmit={handleOnSubmit}